Add example for bootstrap-chat renderResponse, fix exports, misspelling, and usage
